Greg's todo changes #1
I've added a few things to the todo list, mostly for myself, and to get to the github stuff. I am not totally sure but I believe the the bottom section was for questions that need to be answered? so I added one that needs to be answered. Tell me what I need to do better.